Add internal sqliteutil package.
